backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
firebase-admin
gunicorn
cloudinary

google-generativeai
pinecone
requests
pypdf
